refactor: 거래내역 타입분리 (#26)
