+++ client/src/type/transaction.ts
@@ -1,0 +1,13 @@
+type PaymentType = {
+  category: string;
+  title: string;
+  method: string;
+  price: number;
+};
+
+export type DayTransactionType = {
+  date: string;
+  transaction: Array<PaymentType>;
+};
+
+export type TransactionType = Array<DayTransactionType>;
