+++ client/src/type/type.ts
@@ -1,0 +1,3 @@
+export type objType = {
+  [key: string]: any;
+};
